+++ src/combinators/dedupeWith.test.js
@@ -1,69 +1,55 @@
 import { eq } from 'fkit'
 
-import Signal from '../Signal'
 import dedupeWith from './dedupeWith'
+import { mockSignal } from '../emitter'
 
+let s
 let valueSpy, errorSpy, completeSpy
 
 describe('dedupeWith', () => {
   beforeEach(() => {
+    s = mockSignal()
+
     valueSpy = jest.fn()
     errorSpy = jest.fn()
     completeSpy = jest.fn()
   })
 
   it('removes duplicate signal values with a comparator function', () => {
-    let value
-    const s = new Signal(emit => {
-      value = emit.value
-    })
-
     dedupeWith(eq, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(valueSpy).not.toHaveBeenCalled()
-    value('foo')
-    value('foo')
+    s.value('foo')
+    s.value('foo')
     expect(valueSpy).toHaveBeenCalledTimes(1)
     expect(valueSpy).toHaveBeenLastCalledWith('foo')
-    value('bar')
+    s.value('bar')
     expect(valueSpy).toHaveBeenCalledTimes(2)
     expect(valueSpy).toHaveBeenLastCalledWith('bar')
   })
 
   it('emits an error when the given signal emits an error', () => {
-    let error
-    const s = new Signal(emit => {
-      error = emit.error
-    })
-
     dedupeWith(eq, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(errorSpy).not.toHaveBeenCalled()
-    error('foo')
+    s.error('foo')
     expect(errorSpy).toHaveBeenCalledTimes(1)
     expect(errorSpy).toHaveBeenCalledWith('foo')
   })
 
   it('completes when the given signal is completed', () => {
-    let complete
-    const s = new Signal(emit => {
-      complete = emit.complete
-    })
-
     dedupeWith(eq, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(completeSpy).not.toHaveBeenCalled()
-    complete()
+    s.complete()
     expect(completeSpy).toHaveBeenCalledTimes(1)
   })
 
   it('unmounts the given signal when the returned signal is unsubscribed', () => {
-    const unmount = jest.fn()
-    const s = new Signal(() => unmount)
     const a = dedupeWith(eq, s).subscribe()
 
-    expect(unmount).not.toHaveBeenCalled()
+    expect(s.unmount).not.toHaveBeenCalled()
     a.unsubscribe()
-    expect(unmount).toHaveBeenCalledTimes(1)
+    expect(s.unmount).toHaveBeenCalledTimes(1)
   })
 })