+++ src/combinators/endWith.test.js
@@ -1,42 +1,35 @@
-import delay from './delay'
+import endWith from './endWith'
 import { mockSignal } from '../emitter'
 
 let s
 let valueSpy, errorSpy, completeSpy
 
-describe('delay', () => {
+describe('endWith', () => {
   beforeEach(() => {
     s = mockSignal()
 
     valueSpy = jest.fn()
     errorSpy = jest.fn()
     completeSpy = jest.fn()
-    jest.useFakeTimers()
   })
 
-  afterEach(() => {
-    jest.useRealTimers()
-  })
+  it('emits the given value after all other values', () => {
+    endWith(3, s).subscribe(valueSpy, errorSpy, completeSpy)
 
-  it('delays the signal values', () => {
-    delay(1000, s).subscribe(valueSpy, errorSpy, completeSpy)
-
-    s.value(1)
-    s.value(2)
-    jest.advanceTimersByTime(500)
-    s.value(3)
     expect(valueSpy).not.toHaveBeenCalled()
-    jest.advanceTimersByTime(500)
+    s.value(1)
+    expect(valueSpy).toHaveBeenCalledTimes(1)
+    expect(valueSpy).toHaveBeenLastCalledWith(1)
+    s.value(2)
     expect(valueSpy).toHaveBeenCalledTimes(2)
-    expect(valueSpy).toHaveBeenNthCalledWith(1, 1)
-    expect(valueSpy).toHaveBeenNthCalledWith(2, 2)
-    jest.advanceTimersByTime(500)
+    expect(valueSpy).toHaveBeenLastCalledWith(2)
+    s.complete()
     expect(valueSpy).toHaveBeenCalledTimes(3)
-    expect(valueSpy).toHaveBeenNthCalledWith(3, 3)
+    expect(valueSpy).toHaveBeenLastCalledWith(3)
   })
 
   it('emits an error when the given signal emits an error', () => {
-    delay(1000, s).subscribe(valueSpy, errorSpy, completeSpy)
+    endWith(0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(errorSpy).not.toHaveBeenCalled()
     s.error('foo')
@@ -45,7 +38,7 @@ describe('delay', () => {
   })
 
   it('completes when the given signal is completed', () => {
-    delay(1000, s).subscribe(valueSpy, errorSpy, completeSpy)
+    endWith(0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(completeSpy).not.toHaveBeenCalled()
     s.complete()
@@ -53,7 +46,7 @@ describe('delay', () => {
   })
 
   it('unmounts the given signal when the returned signal is unsubscribed', () => {
-    const a = delay(1000, s).subscribe()
+    const a = endWith(0, s).subscribe()
 
     expect(s.unmount).not.toHaveBeenCalled()
     a.unsubscribe()