+++ src/combinators/fold.test.js
@@ -1,69 +1,54 @@
 import { add, always } from 'fkit'
 
-import Signal from '../Signal'
 import fold from './fold'
+import { mockSignal } from '../emitter'
 
+let s
 let valueSpy, errorSpy, completeSpy
 
 describe('fold', () => {
   beforeEach(() => {
+    s = mockSignal()
+
     valueSpy = jest.fn()
     errorSpy = jest.fn()
     completeSpy = jest.fn()
   })
 
   it('folds a function over the signal values', () => {
-    let value, complete
-    const s = new Signal(emit => {
-      value = emit.value
-      complete = emit.complete
-    })
-
     fold(add, 0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
-    value(1)
-    value(2)
-    value(3)
+    s.value(1)
+    s.value(2)
+    s.value(3)
     expect(valueSpy).not.toHaveBeenCalled()
-    complete(1)
+    s.complete(1)
     expect(valueSpy).toHaveBeenCalledTimes(1)
     expect(valueSpy).toHaveBeenCalledWith(6)
   })
 
   it('emits an error when the given signal emits an error', () => {
-    let error
-    const s = new Signal(emit => {
-      error = emit.error
-    })
-
     fold(always(), 0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(errorSpy).not.toHaveBeenCalled()
-    error('foo')
+    s.error('foo')
     expect(errorSpy).toHaveBeenCalledTimes(1)
     expect(errorSpy).toHaveBeenCalledWith('foo')
   })
 
   it('completes when the given signal is completed', () => {
-    let complete
-    const s = new Signal(emit => {
-      complete = emit.complete
-    })
-
     fold(always(), 0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(completeSpy).not.toHaveBeenCalled()
-    complete()
+    s.complete()
     expect(completeSpy).toHaveBeenCalledTimes(1)
   })
 
   it('unmounts the given signal when the returned signal is unsubscribed', () => {
-    const unmount = jest.fn()
-    const s = new Signal(() => unmount)
     const a = fold(always(), 0, s).subscribe()
 
-    expect(unmount).not.toHaveBeenCalled()
+    expect(s.unmount).not.toHaveBeenCalled()
     a.unsubscribe()
-    expect(unmount).toHaveBeenCalledTimes(1)
+    expect(s.unmount).toHaveBeenCalledTimes(1)
   })
 })