+++ src/combinators/scan.test.js
@@ -1,72 +1,67 @@
 import { add, always } from 'fkit'
 
-import Signal from '../Signal'
 import scan from './scan'
+import { asap } from '../scheduler'
+import { mockSignal } from '../emitter'
 
+jest.mock('../scheduler')
+
+let s
 let valueSpy, errorSpy, completeSpy
 
 describe('scan', () => {
   beforeEach(() => {
+    s = mockSignal()
+
     valueSpy = jest.fn()
     errorSpy = jest.fn()
     completeSpy = jest.fn()
+
+    asap.mockImplementation(f => f())
+  })
+
+  afterEach(() => {
+    asap.mockRestore()
   })
 
   it('scans a function over the signal values', () => {
-    let value
-    const s = new Signal(emit => {
-      value = emit.value
-    })
-
     scan(add, 0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(valueSpy).toHaveBeenCalledTimes(1)
     expect(valueSpy).toHaveBeenLastCalledWith(0)
-    value(1)
+    s.value(1)
     expect(valueSpy).toHaveBeenCalledTimes(2)
     expect(valueSpy).toHaveBeenLastCalledWith(1)
-    value(2)
+    s.value(2)
     expect(valueSpy).toHaveBeenCalledTimes(3)
     expect(valueSpy).toHaveBeenLastCalledWith(3)
-    value(3)
+    s.value(3)
     expect(valueSpy).toHaveBeenCalledTimes(4)
     expect(valueSpy).toHaveBeenLastCalledWith(6)
   })
 
   it('emits an error when the given signal emits an error', () => {
-    let error
-    const s = new Signal(emit => {
-      error = emit.error
-    })
-
     scan(always(), 0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(errorSpy).not.toHaveBeenCalled()
-    error('foo')
+    s.error('foo')
     expect(errorSpy).toHaveBeenCalledTimes(1)
     expect(errorSpy).toHaveBeenCalledWith('foo')
   })
 
   it('completes when the given signal is completed', () => {
-    let complete
-    const s = new Signal(emit => {
-      complete = emit.complete
-    })
-
     scan(always(), 0, s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(completeSpy).not.toHaveBeenCalled()
-    complete()
+    s.complete()
     expect(completeSpy).toHaveBeenCalledTimes(1)
   })
 
   it('unmounts the given signal when the returned signal is unsubscribed', () => {
-    const unmount = jest.fn()
-    const s = new Signal(() => unmount)
     const a = scan(always(), 0, s).subscribe()
 
-    expect(unmount).not.toHaveBeenCalled()
+    expect(s.unmount).not.toHaveBeenCalled()
     a.unsubscribe()
-    expect(unmount).toHaveBeenCalledTimes(1)
+    expect(s.unmount).toHaveBeenCalledTimes(1)
   })
 })