+++ src/combinators/sequential.test.js
@@ -1,42 +1,33 @@
-import delay from './delay'
+import { range } from 'fkit'
+
+import sequential from './sequential'
 import { mockSignal } from '../emitter'
 
 let s
 let valueSpy, errorSpy, completeSpy
 
-describe('delay', () => {
+describe('sequential', () => {
   beforeEach(() => {
     s = mockSignal()
 
     valueSpy = jest.fn()
     errorSpy = jest.fn()
     completeSpy = jest.fn()
-    jest.useFakeTimers()
   })
 
-  afterEach(() => {
-    jest.useRealTimers()
-  })
+  it('sequentially emits the values of an array', () => {
+    const t = sequential(range(1, 3), s)
 
-  it('delays the signal values', () => {
-    delay(1000, s).subscribe(valueSpy, errorSpy, completeSpy)
+    t.subscribe(valueSpy, errorSpy, completeSpy)
 
-    s.value(1)
-    s.value(2)
-    jest.advanceTimersByTime(500)
-    s.value(3)
-    expect(valueSpy).not.toHaveBeenCalled()
-    jest.advanceTimersByTime(500)
-    expect(valueSpy).toHaveBeenCalledTimes(2)
-    expect(valueSpy).toHaveBeenNthCalledWith(1, 1)
-    expect(valueSpy).toHaveBeenNthCalledWith(2, 2)
-    jest.advanceTimersByTime(500)
-    expect(valueSpy).toHaveBeenCalledTimes(3)
-    expect(valueSpy).toHaveBeenNthCalledWith(3, 3)
+    range(1, 3).forEach(n => {
+      s.value()
+      expect(valueSpy).toHaveBeenNthCalledWith(n, n)
+    })
   })
 
   it('emits an error when the given signal emits an error', () => {
-    delay(1000, s).subscribe(valueSpy, errorSpy, completeSpy)
+    sequential([], s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(errorSpy).not.toHaveBeenCalled()
     s.error('foo')
@@ -45,7 +36,7 @@ describe('delay', () => {
   })
 
   it('completes when the given signal is completed', () => {
-    delay(1000, s).subscribe(valueSpy, errorSpy, completeSpy)
+    sequential([], s).subscribe(valueSpy, errorSpy, completeSpy)
 
     expect(completeSpy).not.toHaveBeenCalled()
     s.complete()
@@ -53,7 +44,7 @@ describe('delay', () => {
   })
 
   it('unmounts the given signal when the returned signal is unsubscribed', () => {
-    const a = delay(1000, s).subscribe()
+    const a = sequential([], s).subscribe()
 
     expect(s.unmount).not.toHaveBeenCalled()
     a.unsubscribe()