+++ src/emitter.js
@@ -1,3 +1,4 @@
+import Signal from './Signal'
 import events from 'events'
 
 /**
@@ -10,4 +11,20 @@ export default function emitter () {
   emitter.addEventListener = emitter.on
   emitter.removeEventListener = jest.fn()
   return emitter
+}
+
+export function mockSignal () {
+  let value, error, complete
+  const unmount = jest.fn()
+  const s = new Signal(emit => {
+    value = emit.value
+    error = emit.error
+    complete = emit.complete
+    return () => unmount
+  })
+  s.value = a => { value && value(a) }
+  s.error = e => { error && error(e) }
+  s.complete = () => { complete && complete() }
+  s.unmount = unmount
+  return s
 }