+++ src/scheduler.js
@@ -1,0 +1,8 @@
+/**
+ * Executes the function `f` asynchronously.
+ *
+ * @private
+ */
+export function asap (f) {
+  setTimeout(f, 0)
+}