+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 * Add missing combinators to exports
 * Add `concat` function
 * Add `Signal#concat` method
+* Add `always` function
+* Add `cycle` function
+* Add `sequential` function
+* Add `startWith` function
+* Add `endWith` function
 
 ## 1.4.0 (2019-01-08)